refactor(platform): api/hexgate_api package
