+++ platform/api/hexgate_api/core/__init__.py
@@ -1,0 +1,1 @@
+"""Shared infrastructure: signing, DB, ClickHouse, WS relay, mail, SPA."""
