+++ platform/api/hexgate_api/deps/__init__.py
@@ -1,0 +1,1 @@
+"""FastAPI dependency functions (auth gates, resource resolvers)."""
