+++ platform/api/tests/conftest.py
@@ -6,7 +6,7 @@ Two session-wide behaviours:
    leave ``HEXGATE_ALLOW_DEV_USER_HEADER`` unset, but every test in this
    package uses the header to assert tenant-isolation and route gating
    without going through a real cookie login. The seam is gated in
-   ``main.require_user`` (which all the cookie-side deps —
+   ``hexgate_api.main.require_user`` (which all the cookie-side deps —
    ``require_org_member``, ``require_org_membership``,
    ``require_project_admin`` — chain through) so a deployment that
    forgets to disable it doesn't accidentally expose identity-asserting
